requirements and extract auth
