+++ flippyflop.py
@@ -1,47 +1,9 @@
-from __future__ import print_function
 from throttle import throttle
-import pickle
-import os.path
-from googleapiclient.discovery import build
-from google_auth_oauthlib.flow import InstalledAppFlow
-from google.auth.transport.requests import Request
 import pandas as pd
 import datetime
 import time
 
-# If modifying these scopes, delete the file token.pickle.
-SCOPES = ["https://www.googleapis.com/auth/spreadsheets"]
-
 # TODO: Use type hinting
-
-# TODO: Extract get_service to seperate module
-
-
-def get_service():
-    creds = None
-    # The file token.pickle stores the user's access and refresh tokens, and is
-    # created automatically when the authorization flow completes for the first
-    # time.
-    if os.path.exists("token.pickle"):
-        with open("token.pickle", "rb") as token:
-            creds = pickle.load(token)
-    # If there are no (valid) credentials available, let the user log in.
-    if not creds or not creds.valid:
-        if creds and creds.expired and creds.refresh_token:
-            creds.refresh(Request())
-        else:
-            flow = InstalledAppFlow.from_client_secrets_file(
-                "credentials.json", SCOPES
-            )
-            creds = flow.run_local_server(port=0)
-        # Save the credentials for the next run
-        with open("token.pickle", "wb") as token:
-            pickle.dump(creds, token)
-
-    service = build("sheets", "v4", credentials=creds)
-
-    return service
-
 
 class FlippyFlop:
     def __init__(self, service, spreadsheet_id):
@@ -158,6 +120,3 @@ class FlippyFlop:
         )
         return result["values"]
 
-
-if __name__ == "__main__":
-    service = get_service()
