+++ requirements.txt
@@ -1,0 +1,6 @@
+pandas
+google-api-python-client 
+google-auth-httplib2 
+google-auth-oauthlib
+pytest
+pytest-freezegun
